the second version
the second version

+++ gl_shop_datas/src/main/java/com/appabc/datas/service/system/impl/UploadImagesServiceImpl.java
@@ -1,17 +1,26 @@
 package com.appabc.datas.service.system.impl;
 
+import java.io.File;
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.commons.lang3.StringUtils;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.appabc.bean.bo.ViewImgsBean;
+import com.appabc.bean.enums.FileInfo.FileOType;
+import com.appabc.bean.enums.FileInfo.FileStyle;
 import com.appabc.bean.pvo.TUploadImages;
 import com.appabc.common.base.QueryContext;
 import com.appabc.common.base.service.BaseService;
+import com.appabc.common.utils.SystemConstant;
 import com.appabc.datas.dao.system.IUploadImagesDAO;
 import com.appabc.datas.service.system.IUploadImagesService;
+import com.appabc.tools.utils.SystemParamsManager;
 
 /**
  * @Description :
@@ -24,116 +33,129 @@ import com.appabc.datas.service.system.IUploadImagesService;
 
 @Service(value = "IUploadImagesService")
 public class UploadImagesServiceImpl extends BaseService<TUploadImages> implements IUploadImagesService {
+	
+	private Logger logger = Logger.getLogger(this.getClass());
 
 	@Autowired
 	private IUploadImagesDAO iUploadImagesDAO;
+	@Autowired
+	private SystemParamsManager spm;
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#add(com.appabc.common.base
 	 * .bean.BaseBean)
 	 */
-	@Override
+
 	public void add(TUploadImages entity) {
 		iUploadImagesDAO.save(entity);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#modify(com.appabc.common.
 	 * base.bean.BaseBean)
 	 */
-	@Override
+
 	public void modify(TUploadImages entity) {
 		iUploadImagesDAO.update(entity);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#delete(com.appabc.common.
 	 * base.bean.BaseBean)
 	 */
-	@Override
+
 	public void delete(TUploadImages entity) {
+		if(entity == null) return;
 		iUploadImagesDAO.delete(entity);
+		if(StringUtils.isNotEmpty(entity.getFpath())){
+			// 企业卸货地址和询单卸货地址图片删除需要判断图片的物理引用关系，存在多于1份引用关系不能删除
+			if(entity.getOtype() != null && (FileOType.FILE_OTYPE_ADDRESS_ORDER.equals(entity.getOtype()) || FileOType.FILE_OTYPE_ADDRESS.equals(entity.getOtype()))){ 
+				int count = this.iUploadImagesDAO.getCountByFpath(entity.getFpath());
+				if(count > 1) return;
+			}
+			delFile(spm.getString(SystemConstant.UPLOADFILE_DIR) + entity.getFpath());
+		}
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#delete(java.io.Serializable)
 	 */
-	@Override
+
 	public void delete(Serializable id) {
 		iUploadImagesDAO.delete(id);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#query(com.appabc.common.base
 	 * .bean.BaseBean)
 	 */
-	@Override
+
 	public TUploadImages query(TUploadImages entity) {
 		return iUploadImagesDAO.query(entity);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#query(java.io.Serializable)
 	 */
-	@Override
+
 	public TUploadImages query(Serializable id) {
 		return iUploadImagesDAO.query(id);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#queryForList(com.appabc.common
 	 * .base.bean.BaseBean)
 	 */
-	@Override
+
 	public List<TUploadImages> queryForList(TUploadImages entity) {
 		return iUploadImagesDAO.queryForList(entity);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#queryForList(java.util.Map)
 	 */
-	@Override
+
 	public List<TUploadImages> queryForList(Map<String, ?> args) {
 		return iUploadImagesDAO.queryForList(args);
 	}
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see
 	 * com.appabc.common.base.service.IBaseService#queryListForPagination(com
 	 * .appabc.common.base.QueryContext)
 	 */
-	@Override
+
 	public QueryContext<TUploadImages> queryListForPagination(
 			QueryContext<TUploadImages> qContext) {
 		return iUploadImagesDAO.queryListForPagination(qContext);
 	}
-	
+
 	/* (non-Javadoc)根据ID获取图片的显示URL
 	 * @see com.appabc.datas.service.file.IUploadImagesService#getUrlByid(int)
 	 */
@@ -157,7 +179,7 @@ public class UploadImagesServiceImpl extends BaseService<TUploadImages> implemen
 			}
 			imgUrls.append(uiList.get(i).getFullpath());
 		}
-		
+
 		return imgUrls.toString();
 	}
 
@@ -167,8 +189,8 @@ public class UploadImagesServiceImpl extends BaseService<TUploadImages> implemen
 	public String getUrlsByOidAndOtype(String oid, String otype) {
 		TUploadImages entity = new TUploadImages();
 		entity.setOid(oid);
-		entity.setOtype(otype);
-		
+		entity.setOtype(FileOType.enumOf(otype));
+
 		return getUrlsByEntity(entity);
 	}
 
@@ -178,20 +200,154 @@ public class UploadImagesServiceImpl extends BaseService<TUploadImages> implemen
 	public void delByOidAndOtype(String oid, String otype) {
 		TUploadImages entity = new TUploadImages();
 		entity.setOid(oid);
-		entity.setOtype(otype);
-		
-		this.iUploadImagesDAO.delete(entity);
+		entity.setOtype(FileOType.enumOf(otype));
+
+		delete(entity);
 	}
 
 	/* (non-Javadoc)获取List,根据业务ID和业务类型获取一组文件信息
 	 * @see com.appabc.datas.service.system.IUploadImagesService#getListByOidAndOtype(java.lang.String, java.lang.String)
 	 */
-	@Override
 	public List<TUploadImages> getListByOidAndOtype(String oid, String otype) {
 		TUploadImages entity = new TUploadImages();
 		entity.setOid(oid);
-		entity.setOtype(otype);
+		entity.setOtype(FileOType.enumOf(otype));
 		return iUploadImagesDAO.queryForList(entity);
+	}
+
+	/* (non-Javadoc)获取图片信息
+	 * @see com.appabc.datas.service.system.IUploadImagesService#getViewImgsByOidAndOtype(java.lang.String, java.lang.String)
+	 */
+	public List<ViewImgsBean> getViewImgsByOidAndOtype(String oid, String otype) {
+
+		List<ViewImgsBean> vImgList = new ArrayList<ViewImgsBean>();
+
+		ViewImgsBean vi = null;
+
+		TUploadImages entity = new TUploadImages();
+		entity.setOid(oid);
+		entity.setOtype(FileOType.enumOf(otype));
+		List<TUploadImages> uiList = iUploadImagesDAO.queryForList(entity);
+
+		for(TUploadImages ui : uiList) {
+			if(ui.getFstyle() == null || ui.getFstyle().equals(FileStyle.FILE_STYLE_ORIGINAL)) { // 先将原图信息放入BEAN
+				vi = new ViewImgsBean();
+				vi.setId(ui.getId());
+				vi.setUrl(ui.getFullpath());
+				vImgList.add(vi);
+			}
+		}
+
+		for(TUploadImages ui : uiList) { // 缩略图信息放入BEAN
+			if(ui.getFstyle() != null && ui.getFstyle().equals(FileStyle.FILE_STYLE_SMALL)) {
+				for(ViewImgsBean vib : vImgList) {
+					if(ui.getPid().equals(Integer.parseInt(vib.getId())) ){
+						vib.setThumbnailSmall(ui.getFullpath());
+						continue;
+					}
+				}
+			}
+		}
+
+		return vImgList;
+	}
+
+	/* (non-Javadoc)将图片与业务关联
+	 * @see com.appabc.datas.service.system.IUploadImagesService#updateOtypeAndOid(java.lang.String, java.lang.String, java.lang.String)
+	 */
+	public void updateOtypeAndOid(String oid, FileOType otype, String fileid) {
+
+		TUploadImages ui = this.iUploadImagesDAO.query(fileid);
+		if(ui != null){
+			ui.setOid(oid);
+			ui.setOtype(otype);
+			this.iUploadImagesDAO.update(ui); // 更新原文件关联
+
+			TUploadImages qEntity = new TUploadImages();
+			qEntity.setPid(Integer.parseInt(fileid));
+			List<TUploadImages> uiList = this.iUploadImagesDAO.queryForList(qEntity); // 查询子文件（略图之类）
+
+			for(TUploadImages uimg : uiList){
+				uimg.setOid(oid);
+				uimg.setOtype(otype);
+				this.iUploadImagesDAO.update(uimg); // 更新子文件关联
+			}
+		}
+
+
+	}
+
+	/* (non-Javadoc)将一批图片与业务关联，并删除不需要关联的旧图片
+	 * @see com.appabc.datas.service.system.IUploadImagesService#updateOtypeAndOidBatch(java.lang.String, com.appabc.bean.enums.FileInfo.FileOType, java.lang.String[])
+	 */
+	public void updateOtypeAndOidBatch(String oid, FileOType otype, String[] fileids) {
+		TUploadImages uiBean = new TUploadImages();
+		uiBean.setOid(oid);
+		uiBean.setOtype(otype);
+		List<TUploadImages> imgList = iUploadImagesDAO.queryForList(uiBean);
+
+		if(fileids != null){
+			for(String imgid : fileids) {
+				boolean isUpdate = true;
+				for(int i=imgList.size()-1; i>=0; i--) { // 过滤不需要更新的文件
+					if(imgList.get(i).getId().equals(imgid) || (imgList.get(i).getPid() != null && imgList.get(i).getPid().equals(Integer.valueOf(imgid)))){
+						imgList.remove(i);
+						isUpdate = false;
+					}
+				}
+				if(isUpdate == true){
+					updateOtypeAndOid(oid, otype, imgid);
+				}
+			}
+		}
+
+		for(int i=0; i<imgList.size(); i++){ // 删除不需要关联的旧图片
+			this.iUploadImagesDAO.delete(imgList.get(i).getId());
+		}
+
+	}
+	
+	/* (non-Javadoc)
+	 * @see com.appabc.datas.service.system.IUploadImagesService#getViewImgsById(java.lang.String)
+	 */
+	@Override
+	public ViewImgsBean getViewImgsById(String imgid) {
+		
+		ViewImgsBean vi = new ViewImgsBean();
+		if(StringUtils.isNotEmpty(imgid)){
+			TUploadImages ui = this.iUploadImagesDAO.query(imgid);
+			if(ui != null){
+				vi.setId(ui.getId());
+				vi.setUrl(ui.getFullpath());
+				
+				// 略图URL
+				TUploadImages uiBean = new TUploadImages();
+				uiBean.setPid(Integer.parseInt(imgid));
+				List<TUploadImages> imgList = iUploadImagesDAO.queryForList(uiBean);
+				if(imgList != null && imgList.size()>0 && imgList.get(0)!=null){
+					vi.setThumbnailSmall(imgList.get(0).getFullpath());
+				}
+			}
+			
+		}
+		
+		return vi;
+	}
+	
+	/**
+	 * 文件物理删除
+	 * @param filePath
+	 */
+	private void delFile(String filePath) {
+		logger.debug("del File path is:" + filePath);
+		try {
+			File file = new File(filePath);
+			if(file.isFile()){
+				file.delete();
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
 	}
 	
 }
